Consolidate the app in one single component
